Reviewed-on: https://dev.bazaari.com.au/sp-core/core/pulls/553

+++ system/Base/Providers/CacheServiceProvider/CacheTools.php
@@ -4,7 +4,7 @@ namespace System\Base\Providers\CacheServiceProvider;
 
 use League\Flysystem\FilesystemException;
 use League\Flysystem\UnableToReadFile;
-use Phalcon\Helper\Json;
+use League\Flysystem\UnableToWriteFile;
 
 class CacheTools
 {
@@ -18,9 +18,13 @@ class CacheTools
 
 	protected $localContent;
 
+	protected $opCache;
+
 	protected $index;
 
-	public function __construct($cacheConfig, array $caches, $localContent)
+	public $caches;
+
+	public function __construct($cacheConfig, array $caches, $localContent, $opCache)
 	{
 		$this->cacheConfig = $cacheConfig;
 
@@ -32,9 +36,18 @@ class CacheTools
 			$this->cacheService = 'streamCache';
 		}
 
+		$this->caches = $caches;
+
 		$this->cache = $caches[$this->cacheService];
 
 		$this->localContent = $localContent;
+
+		$this->opCache = $opCache;
+	}
+
+	public function getAvailableCaches()
+	{
+		return array_keys($this->caches);
 	}
 
 	public function addModelCacheParameters($parameters = null, $cacheName = null)
@@ -69,7 +82,7 @@ class CacheTools
 
 	public function generateCacheKey($cacheKey)
 	{
-		$cacheKey = Json::encode($cacheKey);
+		$cacheKey = $this->helper->encode($cacheKey);
 
 		$cacheKey = preg_replace('/[^A-Za-z0-9.-]/', '', $cacheKey);
 
@@ -113,16 +126,26 @@ class CacheTools
 
 	public function createIndexFile($cacheName, $recreate = false)
 	{
-		if (!$this->localContent->fileExists('var/storage/cache/stream/index/' . $cacheName . '.json') || $recreate) {
-			$content['all'] = [];
-			$content['list'] = [];
-			$content['ids'] = [];
+		$content['all'] = [];
+		$content['list'] = [];
+		$content['ids'] = [];
 
-			$this->localContent->write('var/storage/cache/stream/index/' . $cacheName . '.json', Json::encode($content));
+		if ($this->opCache) {
+			if (!$this->opCache->getCache($cacheName, 'index')) {
+				$this->opCache->setCache($cacheName, $content, 'index');
+			} else if ($recreate) {
+				$this->opCache->resetCache($cacheName, $content, 'index');
+			}
+		} else {
+			if (!$this->localContent->fileExists('var/storage/cache/stream/index/' . $cacheName . '.json') ||
+				$recreate
+			) {
+				$this->localContent->write('var/storage/cache/stream/index/' . $cacheName . '.json', $this->helper->encode($content));
+			}
 		}
 	}
 
-	public function updateIndex($cacheName, $parameters, $list = false, $id = false, $object)
+	public function updateIndex($cacheName, $parameters, $object, $list = false, $id = false)
 	{
 		if (!isset($parameters['cache']) && !isset($parameters['cache']['key'])) {
 			return;
@@ -157,8 +180,20 @@ class CacheTools
 				}
 			}
 
-			$this->localContent->write('var/storage/cache/stream/index/' . $cacheName . '.json', Json::encode($index));
+			if ($this->opCache) {
+				$this->opCache->resetCache($cacheName, $index, 'index');
+			} else {
+				try {
+					$this->localContent->write('var/storage/cache/stream/index/' . $cacheName . '.json', $this->helper->encode($index));
+				} catch (FilesystemException | UnableToWriteFile | \Exception $e) {
+					return false;
+				}
+			}
+
+			return true;
 		}
+
+		return false;
 	}
 
 	public function resetCache($cacheName = null, $id = null, $removeId = false)
@@ -168,7 +203,11 @@ class CacheTools
 		if (!$cacheName) {//Only do this in maintenance mode
 			$this->cache->clear();
 
-			$this->localContent->deleteDirectory('var/storage/cache/stream/');
+			if ($this->opCache) {
+				$this->opCache->resetCache(null, 'index');
+			} else {
+				$this->localContent->deleteDirectory('var/storage/cache/stream/');
+			}
 
 			$this->index = null;
 
@@ -177,27 +216,25 @@ class CacheTools
 
 		if ($cacheName && $id) {
 			$keys = $this->getKeysFromIndex($cacheName, $id);
-
-			if ($keys && is_array($keys) && count($keys) > 0) {
-				foreach ($keys as $key => $parameters) {
-					$this->deleteCache($key);
-
-					$this->removeKeyFromIndex($cacheName, $key, $id, $removeId);
-				}
-
-				$this->localContent->write('var/storage/cache/stream/index/' . $cacheName . '.json', Json::encode($this->index));
-			}
 		} else {
 			$keys = $this->getKeysFromIndex($cacheName);
+		}
 
-			if ($keys && is_array($keys) && count($keys) > 0) {
-				foreach ($keys as $key => $parameters) {
-					$this->deleteCache($key);
+		if ($keys && is_array($keys) && count($keys) > 0) {
+			foreach ($keys as $key => $parameters) {
+				$this->deleteCache($key);
 
+				if ($cacheName && $id) {
+					$this->removeKeyFromIndex($cacheName, $key, $id, $removeId);
+				} else {
 					$this->removeKeyFromIndex($cacheName, $key);
 				}
+			}
 
-				$this->localContent->write('var/storage/cache/stream/index/' . $cacheName . '.json', Json::encode($this->index));
+			if ($this->opCache) {
+				$this->opCache->resetCache($cacheName, $this->index, 'index');
+			} else {
+				$this->localContent->write('var/storage/cache/stream/index/' . $cacheName . '.json', $this->helper->encode($this->index));
 			}
 		}
 
@@ -249,24 +286,32 @@ class CacheTools
 
 	protected function getIndex($cacheName)
 	{
-		try {
-			$index = $this->localContent->read('var/storage/cache/stream/index/' . $cacheName . '.json');
-
-			$index = Json::decode($index, true);
+		if ($this->opCache) {
+			$index = $this->opCache->getCache($cacheName, 'index');
 
 			$this->index = $index;
 
 			return $index;
-		} catch (FilesystemException | UnableToReadFile | \Exception $e) {
-			if (str_contains($e->getMessage(), "json_decode") ||
-				get_class($e) !== 'FilesystemException' ||
-				get_class($e) !== 'UnableToReadFile'
-			) {
-				$this->createIndexFile($cacheName, true);
+		} else {
+			try {
+				$index = $this->localContent->read('var/storage/cache/stream/index/' . $cacheName . '.json');
 
-				return $this->getIndex($cacheName);
-			} else {
-				throw $e;
+				$index = $this->helper->decode($index, true);
+
+				$this->index = $index;
+
+				return $index;
+			} catch (FilesystemException | UnableToReadFile | \Exception $e) {
+				if (str_contains($e->getMessage(), "json_decode") ||
+					get_class($e) !== 'FilesystemException' ||
+					get_class($e) !== 'UnableToReadFile'
+				) {
+					$this->createIndexFile($cacheName, true);
+
+					return $this->getIndex($cacheName);
+				} else {
+					throw $e;
+				}
 			}
 		}
 	}
